requirement.txt
Django==1.11.12
djangorestframework==3.7.7
pytz==2018.3
